add DialogAlert

+++ feature/master/main/src/main/java/com/hfad/main/presentation/employeepack/EditEmployeeScreen.kt
@@ -15,11 +15,16 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -51,17 +56,14 @@ fun EditEmployeeScreen(
 ) {
     val state = viewModel.state
     val backgroundColor = MaterialTheme.colorScheme.background
-    val selectedImageUri = rememberSaveable {
-        mutableStateOf<Uri?>(null)
-    }
+    val placeholderImage = painterResource(id = R.drawable.employeeorc)
+    val selectedImageUri = rememberSaveable { mutableStateOf<Uri?>(null) }
+    var showDeleteDialog by remember { mutableStateOf(false) }
     val imageLauncher = rememberLauncherForActivityResult(
         contract = ActivityResultContracts.GetContent()
     ) { uri ->
         selectedImageUri.value = uri
     }
-    val placeholderImage = painterResource(
-        id = R.drawable.employeeorc
-    )
 
     ObserveAsEvents(flow = viewModel.navigationEventsChannelFlow) { event ->
         when (event) {
@@ -144,9 +146,30 @@ fun EditEmployeeScreen(
         Spacer(modifier = Modifier.height(10.dp))
         FzRedOutlinedButton(
             onClick = {
-                viewModel.onDeleteAccountClick()
+                showDeleteDialog = true
             },
             text = { Text(text = stringResource(com.hfad.main.R.string.DeleteAccount)) }
+        )
+    }
+
+    if (showDeleteDialog) {
+        AlertDialog(
+            onDismissRequest = { showDeleteDialog = false },
+            title = { Text(text = stringResource(com.hfad.main.R.string.ConfirmDelete)) },
+            text = { Text(text = stringResource(com.hfad.main.R.string.SureDeleteAccount)) },
+            confirmButton = {
+                Button(onClick = {
+                    viewModel.onDeleteAccountClick()
+                    showDeleteDialog = false
+                }) {
+                    Text(text = stringResource(com.hfad.main.R.string.Yes))
+                }
+            },
+            dismissButton = {
+                Button(onClick = { showDeleteDialog = false }) {
+                    Text(text = stringResource(com.hfad.main.R.string.No))
+                }
+            }
         )
     }
 }
